Cleanup + Minestom Visualiser

+++ src/main/java/pl/pateman/dynamicaabbtree/AreaAABBHeuristicFunction.java
@@ -5,20 +5,17 @@ import org.joml.AABBf;
 import static pl.pateman.dynamicaabbtree.AABBUtils.getArea;
 
 
-public class AreaAABBHeuristicFunction<T extends Boundable> implements AABBTreeHeuristicFunction<T>
-{
-   private final AABBf temp;
+public class AreaAABBHeuristicFunction<T extends Boundable> implements AABBTreeHeuristicFunction<T> {
+    private final AABBf temp;
 
-   public AreaAABBHeuristicFunction()
-   {
-      temp = new AABBf();
-   }
+    public AreaAABBHeuristicFunction() {
+        temp = new AABBf();
+    }
 
-   @Override
-   public HeuristicResult getInsertionHeuristic(AABBf left, AABBf right, T object, AABBf objectAABB)
-   {
-      float diffA = getArea(left.union(objectAABB, temp)) - getArea(left);
-      float diffB = getArea(right.union(objectAABB, temp)) - getArea(right);
-      return diffA < diffB ? HeuristicResult.LEFT : HeuristicResult.RIGHT;
-   }
+    @Override
+    public HeuristicResult getInsertionHeuristic(AABBf left, AABBf right, T object, AABBf objectAABB) {
+        double diffA = getArea(left.union(objectAABB, temp)) - getArea(left);
+        double diffB = getArea(right.union(objectAABB, temp)) - getArea(right);
+        return diffA < diffB ? HeuristicResult.LEFT : HeuristicResult.RIGHT;
+    }
 }
